Clear up some workflow confusion
The "package" version was being assigned to the "asl3-ver" input and "ASL3_VER"
env var.  This resulted in some confusion with our having a GitHub "ASL3"
repository.  We're not going to use "package-ver" and PACKAGE_VER"!

Also, we were cloning the "ASL3" repository into the docker build tree but not
using it in any way.  We will no longer pull in this repo.