Move to Asterisk-22.x.x LTS
These changes include a patch to re-add detection of the ALSA library to
the build system.  This is needed for chan_simpleusb and chan_usbradio.

Pulled (from ASL3-Manual) and updated documentation on using phreaknet to
build asterisk + app_rpt.  Will be updating the ASL3 Manual content to
point back to this repo.