Correct "astmoddir" path
Our template `asterisk.conf` file includes an "astmoddir" variable that
should reference the Asterisk module directory.  On Debian, the path is
architecture specific.  So far, despite the path being incorrect on some
architectures, we've been OK as `asterisk` knows were to find/load it's
own modules.  But, the Asterisk "testsuite" is not as smart and uses this
variable to find the module directory.  This change ensures that the path
is correct.